requirements.txt
pandas>=2.2.0
aiohttp>=3.10.0
python-dotenv>=1.0.0
requests>=2.32.0